opensearch_project/requirements.txt
opensearch-py
python-dotenv
pydantic
fastapi
uvicorn
faker
pytest
requests
python-dotenv
